[Pipelines] Let Llava work with no images as input
Split Pixtral into two MAX graphs:
1- vision enocder and projection layer into the embedding space of
languag model.

2- language model (mistral)

It passes the output of the vision encoder to the decoder.

I kept the original layers and infra that uses one graph for Pixtral to
be available for use whenever we have mo.if

Tested it with multiple image and text prompots and text-only prompots
on Pixtral.

MODULAR_ORIG_COMMIT_REV_ID: 100f670162c5e880bd279518aecd22327609bfc3