[Graph API][Examples] Improve Mandelbrot ****** example and simplify graph creation in addition example
The Mandelbrot set custom operation example was creating tensors for
initial values of the complex number C on the host and then passing them
into the graph operation. This reworks the operation and example to
instead perform the initialization of C inside the Mojo operation and
avoid any copies into the operation. This is useful as an example,
because it shows how to handle graph and operation with no tensor
inputs.

This also follows on from comments on the initial implementation of the
"hello world" addition example and changes that to use a simpler version
of Python graph creation when dealing with a single operation. That
example now uses explicit Tensor creation and movement to the
accelerator device, due to the implicit copying of tensors being
confusing for some people who tried out this example.

MODULAR_ORIG_COMMIT_REV_ID: 7a33f4a149169712e4aad50a43ff7b2f75db0fb3