[Pipelines] Yield to the event loop during generation (#55290)
Without this, in practice the entire generation is synchronous, and
there is no time for asyncio to interject and tell us we have been
interrupted. As a result, without this, it requires two SIGINTs to
terminate a generation. With this, we have an opportunity to be
interrupted after each generated token, and a single SIGINT is
sufficient.

Fixes AITLIB-132.

MODULAR_ORIG_COMMIT_REV_ID: 39c9cb5e7fee604df244b39b0cbd75aea0a00d23