[KVCache] Make prefix caching blocks eviction run in O(1) time (#55631)
This PR achieves this by mantaining a LRU cache (aka an Ordered Dict)
that contains all of the evictable nodes in the Trie. Each item in the
cache consists of a unique node id and a Trie node.

Trie nodes are inserted into the LRU cache upon creation (as ref count
is zero), and when sequences mark nodes as unused. When inserted, they
are moved to the back of the LRU cache via `move_to_end`. Moving to end
is called implictly whenever we write `lru_cache[key] = node`, even when
the key is already in the cache.

When evicting, we pop a node from the from of the LRU cache. We can
evict some number of blocks from the node. If we only need to evict a
portion of the blocks, we can reinsert the node in the front of the
cache (`move_to_front`).

We guarantee that the LRU leaf nodes are evicted first. In the LRU
cache, we guarantee that the parent node will only be evicted after all
of its children.

The current prefix caching benchmarks on A100 result in little to no
block evictions occuring. Thus, this should have negligible impact on
our perf. However, this will be more important on gpus with smaller VRAM
capacity or in cases where KV Cache capacity is strained.

Fixes E2EOPT-37

MODULAR_ORIG_COMMIT_REV_ID: 5297f9e83b373765003a74aa6bb2f5ef01539690