[MAX] Add Phi3ForCausalLM model support (#55954)
As mentioned in

[Internal link]
there are minor differences between llama and phi4 with the main one
being that the qkv and gate/up proj weights are stacked. Thus, one needs
to handle the attention mechnism and MLP blocks in a special way. We can
query if the keys are present and dispatch for the Phi model.

Note that this does not apply the changes to the quantization and
multi-gpu pathway. I'll do that in a followup if we can find a model
which requires this. In the meantime we probably need to guard in the
reigstery that quantization / multi-gpu is not supported for some archs.

Tested via

```
./bazelw run --config=disable-mypy  SDK/lib/API/python/max/entrypoints/pipelines -- generate --huggingface-repo-id=microsoft/phi-4 --prompt "Write bubble sort in python"  --quantization-encoding bfloat16
```

which produces

```
Write bubble sort in python to sort a list of numbers in ascending order.

[Response]:
def bubble_sort(lst):
    n = len(lst)
    for i in range(n-1):
        for j in range(n-i-1):
            if lst[j] > lst[j+1]:
                lst[j], lst[j+1] = lst[j+1], lst[j]
    return lst<|endoftext|>
```

MODULAR_ORIG_COMMIT_REV_ID: 227f7aeb60f7cef8cdc7ddeadc814e5d351ec565