[MAX] Add OlmoForCausalLM support (#55962)
As described in

[Internal link]
there are minor differences between llama nd olmo with the major being
the use of layernorm instead of rmsnorm as well as qkv clamping logic.

With this I get

```
./bazelw run --config=disable-mypy  SDK/lib/API/python/max/entrypoints/pipelines -- generate --huggingface-repo-id=allenai/OLMo-1B-0724-hf --prompt "Write bubble sort in python"
```

which outputs

```
Bubble sort is a simple sorting algorithm that works by repeatedly swapping adjacent elements. It is a simple and efficient algorithm that is used to sort an array of numbers.

The algorithm works by repeatedly swapping adjacent elements until the array is sorted.

The algorithm is simple and efficient, and it is often used to sort small arrays of numbers.

The algorithm works by repeatedly swapping adjacent elements until the array is sorted.

The algorithm is simple and efficient,
```

Note that the quantization / distributed pathway does not currently work
for this model.

MODULAR_ORIG_COMMIT_REV_ID: fccabd0aa5da779ff97073a09b322321879e4229