[Pipelines] Fix default max-steps option breaking padded CPU inference (#55994)
This changes our behavior to have a "not set" state for the num_steps
setting. If we find ourselves with this not set, then we set it to the
relevant default for the given setup. If it _is_ set, then we validate
that what the user has provided is valid.

MODULAR_ORIG_COMMIT_REV_ID: 2da17647b55211db6a8340247d0fbc998e51ee98